Rework Docker containers and install scripts
Install scripts and containers were modified to build
and use a "bootstrap" compiler for building the exaworks
software.  This is necessary because dependencies shared
between various components of exaworks will induce multiple
rebuilds of the gcc compiler, which takes a very long time.
This is avoided by bootstrapping gcc with the native
compiler, and then rebuilding the newer gcc with itself
in a "base" environment.  In this way, gcc is built twice
for the base environment, but subsequent environments
using the "base" environment will not rebuild gcc.

This commit also adds CentOS7 containers for the
multi-stage build of exaworks that breaks it into
smaller pieces with shorter build times. The Github
Actions workflow was updated to add CentOS containers.