Consolidate Dockerfiles for ExaWorks constituents (#21)
* Consolidate Dockerfiles for ExaWorks constituents

There were Dockerfiles for Ubuntu20 and CentOS7 builds
of ExaWorks and its constituents.  The Dockerfiles
for the base environment were significantly different
but the ones for the constituent pieces only differed
in the initial FROM statement at the top.  This commit
removes the duplicate Dockerfiles for the constituent
parts and uses build arguments (e.g. ARG) to specify the
OS to use at build time.  This elminates 5 extraneous
Dockerfiles and increases maintainability.

* Fix typo in GHA workflow for CentOS container

* Use environment variable in CI workflow to control OS used for containers

* Use env context to access container os variable