Rework install scripts to use pip for Parsl and RADICAL
This allows the latest versions of Parsl and RADICAL to be
installed via pip with miniconda3.  It works around the
problem that Spack packages are not kept up to date for
Parsl and RADICAL.  It also drastically speeds up the
installation process by eliminating extremely long builds
of Parsl and RADICAL dependencies.