Fix multi-platform CI build
Multi-platform is not supported for the "load", only "push".
So, the platform was removed from the initial builds that
load the cotainer and retained in the builds that push afterward.

The Dockerfile for spack-stack was patched to explicitly
call the setup-env.sh script to work around issues with
spack shell support when running under qemu which is needed
during multi-platform builds.